requirements.txt
google-search-results>=2.4.2
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
click>=8.1.0
